refactoring to allow custom p4-14 converter
