+++ frontends/common/parseInput.h
@@ -1,5 +1,5 @@
 /*
-Copyright 2013-present Barefoot Networks, Inc. 
+Copyright 2013-present Barefoot Networks, Inc.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -18,12 +18,43 @@ limitations under the License.
 #define _FRONTENDS_COMMON_PARSEINPUT_H_
 
 #include "frontends/common/options.h"
+#include "frontends/parsers/parserDriver.h"
+#include "frontends/p4/fromv1.0/converters.h"
+#include "frontends/p4/frontend.h"
+#include "lib/error.h"
+#include "lib/source_file.h"
 
 namespace IR {
 class P4Program;
 }  // namespace IR
 
 namespace P4 {
+
+template <typename Input, typename C = P4V1::Converter>
+static const IR::P4Program*
+parseV1Program(Input& stream, const char* sourceFile, unsigned sourceLine,
+               boost::optional<DebugHook> debugHook = boost::none) {
+    // We load the model before parsing the input file, so that the SourceInfo
+    // in the model comes first.
+    C converter;
+    if (debugHook) converter.addDebugHook(*debugHook);
+    converter.loadModel();
+
+    // Parse.
+    const IR::Node* v1 = V1::V1ParserDriver::parse(stream, sourceFile,
+                                                   sourceLine);
+    if (::errorCount() > 0 || v1 == nullptr)
+        return nullptr;
+
+    // Convert to P4-16.
+    if (Log::verbose())
+        std::cerr << "Converting to P4-16" << std::endl;
+    v1 = v1->apply(converter);
+    if (::errorCount() > 0 || v1 == nullptr)
+        return nullptr;
+    BUG_CHECK(v1->is<IR::P4Program>(), "Conversion returned %1%", v1);
+    return v1->to<IR::P4Program>();
+}
 
 /**
  * Parse P4 source from a file. The filename and language version are specified
@@ -33,7 +64,36 @@ namespace P4 {
  * @return a P4-16 IR tree representing the contents of the given file, or null
  * on failure. If failure occurs, an error will also be reported.
  */
-const IR::P4Program* parseP4File(CompilerOptions& options);
+template <typename C = P4V1::Converter>
+const IR::P4Program* parseP4File(CompilerOptions& options) {
+    BUG_CHECK(&options == &P4CContext::get().options(),
+              "Parsing using options that don't match the current "
+              "compiler context");
+    FILE* in = nullptr;
+    if (options.doNotPreprocess) {
+        in = fopen(options.file, "r");
+        if (in == nullptr) {
+            ::error("%s: No such file or directory.", options.file);
+            return nullptr;
+        }
+    } else {
+        in = options.preprocess();
+        if (::errorCount() > 0 || in == nullptr)
+            return nullptr;
+    }
+
+    auto result = options.isv1()
+                ? parseV1Program<FILE*, C>(in, options.file, 1, options.getDebugHook())
+                : P4ParserDriver::parse(in, options.file);
+    options.closeInput(in);
+
+    if (::errorCount() > 0) {
+        ::error("%1% errors encountered, aborting compilation", ::errorCount());
+        return nullptr;
+    }
+    BUG_CHECK(result != nullptr, "Parsing failed, but we didn't report an error");
+    return result;
+}
 
 /**
  * Parse P4 source from the string @input, interpreting it as having language
